pyproject.toml
[project]
name = "newmemory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "sulfastor", email = "torresmoisesa@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "django>=5.1,<6",
    "djangorestframework>=3.16.1",
]

[build-system]
requires = ["uv_build>=0.8.19,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ruff>=0.14.2",
]

[tool.ruff]
target-version = "py312"
line-length = 100
lint.select = ["E", "F", "I", "B", "UP", "C90"]
extend-exclude = ["migrations"]

[tool.ruff.lint.isort]
known-first-party = ["api", "config"]
